backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
pycparser==2.22
pydantic==2.12.5
pydantic-core==2.41.5
PyJWT==2.10.1
python-multipart==0.0.9
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing-inspection==0.4.2
typing_extensions==4.14.1
uvicorn==0.32.1
python-dotenv==1.0.1
opencv-python==4.10.0.84
numpy==1.26.4
pillow==10.4.0